backend/requirements.txt
twilio
python-dotenv
fastapi
uvicorn
python-socketio
httpx
scikit-learn
joblib
pandas
numpy
